fix: add Windows ao installer
